Move code from retroarch.c to video_shader_parse.c
